refactor the src code to focus on flower related python scripts in demo
