--- bin/down.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-#
-# Stops and removes docker containers and images
-
-source bin/setup_environment.sh
-
-docker-compose down "$@"
