+++ bin/stop_server.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+#
+# Stop and remove the container and image associated with the flower server
+
+source bin/setup_environment.sh
+
+docker-compose down flower
