--- bin/test.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-#
-# Run unit tests
-
-source bin/setup_environment.sh
-
-docker-compose run --rm --entrypoint python "$SERVICE" -m pytest -p no:warnings "${@:-tests/}"
-
