--- bin/up.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-#
-# Start docker services
-
-source bin/setup_environment.sh
-
-docker-compose up "$@"
